Developing harvest

+++ db/migrate/20191208205953_create_phs.rb
@@ -1,7 +1,7 @@
 class CreatePhs < ActiveRecord::Migration[6.0]
   def change
     create_table :phs do |t|
-      t.integer :measure
+      t.float :measure
     end
   end
 end
